requirements.txt
aiofile
aiohttp
lxml
m3u8
pycryptodome